pyproject.toml
[project]
name = "data-hunt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "apscheduler>=3.11.1",
    "fastapi>=0.123.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
    "web3>=7.14.0",
    "sqlalchemy>=2.0.45",
    "psycopg2-binary>=2.9.10",
    "psycopg2>=2.9.11",
    "pyjwt>=2.10.1",
]
